% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali_mean_ybinary_algm.R
\name{cali_mean_ybinary_algm}
\alias{cali_mean_ybinary_algm}
\title{Calibrate Estimate of Intervention Mean for Binary Outcome}
\usage{
cali_mean_ybinary_algm(i, gamma, mu_u_tr, mu_u_t, mu_y_t, nsim = 4000)
}
\arguments{
\item{i}{Observation index.}

\item{gamma}{Scalar or vector specifying the sensitivity parameters.}

\item{mu_u_tr}{Matrix of conditional confounder means for all observed treatments
with latent variables in columns.}

\item{mu_u_t}{Matrix of conditional confounder means for treatments of interest
with latent variables in columns.}

\item{mu_y_t}{Scalar or vector that contains naive estimates of treatment effects
ignoring confounding.}

\item{nsim}{Number of simulation sample draws.}
}
\value{
Scalar of calibrated intervention mean.
}
\description{
Calibrate Estimate of Intervention Mean for Binary Outcome
}
