% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GaussianT_GaussianY}
\alias{GaussianT_GaussianY}
\title{Dataset with Gaussian Treatments and Outcomes}
\format{
A data frame with eleven variables: one Gaussian outcome, \code{y}, and ten Gaussian
treatments, \code{t1}, \code{t2}, ..., \code{t10}.
}
\source{
For data generating process, see \code{data-raw/Data_Generation.R}.
}
\usage{
GaussianT_GaussianY
}
\description{
A dataset containing Gaussian treatments and outcomes of 10,000 observations.
}
\keyword{datasets}
