% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mice_est_nulltr}
\alias{mice_est_nulltr}
\title{Estimates of genes' effects on mice body weight using null treatments approach from Miao et al. (2020)}
\format{
A data frame with 17 rows and 6 variables:
\describe{
  \item{esti}{mean estimates of genes' treatment effects on mouse body weight}
  \item{X2.5.}{2.5\% percentile of the estimates of genes' treatment effects on mouse body weight}
  \item{X97.5.}{97.5\% percentile of the estimates of genes' treatment effects on mouse body weight}
  \item{X5.}{5\% percentile of the estimates of genes' treatment effects on mouse body weight}
  \item{X95.}{95\% percentile of the estimates of genes' treatment effects on mouse body weight}
  \item{signif}{significance}
  }
}
\source{
\url{https://arxiv.org/abs/2011.04504}
}
\usage{
mice_est_nulltr
}
\description{
The dataset consists of estimates of treatment effects of 17 genes, which are likely to affect mouse weight,
by using the null treatments approach from Miao et al. (2020),
assuming that at least half of the confounded treatments have no causal effect on the outcome.
}
\keyword{datasets}
