% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{micedata}
\alias{micedata}
\title{Body weight and gene expressions of 287 mice}
\format{
A data frame with forty-three variables: the mice body weight, \code{y},
5 single nucleotide polymorphisms, \code{rs3663003}, \code{rs4136518}, \code{rs3694833},
\code{rs4231406}, \code{rs3661189}, and the rest are thirty-seven genes.
}
\source{
\url{https://arxiv.org/abs/2011.04504}
}
\usage{
micedata
}
\description{
A dataset are collected from 287 mice, including the body weight, 37 gene expressions,
and 5 single nucleotide polymorphisms.
}
\keyword{datasets}
