% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Visualize Estimates of Treatment Effects}
\usage{
plot_estimates(est_df)
}
\arguments{
\item{est_df}{\code{data.frame} containing estimates of treatment effects with estimates' type in columns and
contrasts of interest in rows.}
}
\description{
Visualize Estimates of Treatment Effects
}
\examples{
Please see \code{\link{bcalibrate}} or \code{\link{gcalibrate}}
}
