% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Visualize Estimates of Treatment Effects}
\usage{
plot_estimates(est, show_rv = TRUE)
}
\arguments{
\item{est}{an return object from \code{\link{gcalibrate}} or \code{\link{bcalibrate}}, or
\code{data.frame} containing estimates of treatment effects with estimates' type in columns and
contrasts of interest in rows.}

\item{show_rv}{logical. Whether robustness values should be printed in the plot or not?
Available only for the "worstcase" calibration.}
}
\description{
Visualize Estimates of Treatment Effects
}
\examples{
Please see \code{\link{bcalibrate}} or \code{\link{gcalibrate}}
}
